+++ src/models/queries.js
@@ -1,20 +1,3 @@
-// add user to DB
-function addUser(client, data) {
-    client.query('INSERT INTO users (first_name, last_name, username, email, user_password) VALUES ($1, $2, $3, $4, $5)', [data.firstName, data.lastName, data.username, data.email, data.password])
-}
-
-function returnUser(client) {
-    return new Promise(function (resolve, reject) {
-        client.query('SELECT * FROM users', function (err, response) {
-            if (err) {
-                reject(err)
-            } else {
-                resolve(response.rows)
-            }
-        })
-    })
-}
-
 // add event to DB
 function addEvent(client, data) {
     client.query('INSERT INTO events (event_name, event_description, event_date, event_time) VALUES ($1, $2, $3, $4)', [data.name, data.description, data.date, data.time])
@@ -38,8 +21,6 @@ function addRSVP(client, data) {
 }
 
 module.exports = {
-    'addUser': addUser,
-    'returnUser': returnUser,
     'addEvent': addEvent,
     'returnEvent': returnEvent,
     'addRSVP': addRSVP