+++ changelog.md
@@ -7,9 +7,13 @@ All commits are documented in this file.
  - create event form
  - backend to receive event data from frontend and add it to DB
  - select event from DB and send it to frontend
+ - 3 separate modules for users, events, and rsvps
+ - retrieve(), update(), and delete() for users
 
 ### Changed
  - Date and Time datatypes in event table 
+ - addUser to create()
+ - returnUsers to all()
 
 ### Removed
  - linked_username and linked_event_name columns from RSVP